refactor(component): split component.rb into cohesive concerns
component.rb had grown to ~950 lines mixing four concerns. It is now the
aggregator (module doc + the three public Data value types + includes), and
the behavior lives in:

  * component/registry.rb — the shared inheritance-aware Registry (#115)
  * component/dsl.rb      — the five declaration registries + from_identity
  * component/identity.rb — reactive_token + hot-path ivar precomputation
  * component/helpers.rb  — reply/js, reactive_attrs/root, on/on_client,
    field/select/text bindings, nested-attributes helpers

Pure code motion: every method body, doc comment, constant path
(Component::Action/ComputeDefinition/CollectionDefinition), and visibility
(reactive_token stays private) is verbatim. Zero public API change; the
loader inflects component/dsl.rb -> Component::DSL like the js -> JS
precedent, and the eager-load spec covers the new layout.

Refs #115
